+++ docs/gatsby-config.js
@@ -8,8 +8,9 @@ module.exports = {
         subtitle: 'Apollo Client',
         description: 'A guide to using the Apollo GraphQL Client with React',
         githubRepo: 'apollographql/apollo-client',
-        defaultVersion: 2.5,
+        defaultVersion: 2.6,
         versions: {
+          2.5: 'version-2.5',
           2.4: 'version-2.4',
         },
         checkLinksExceptions: [
@@ -21,12 +22,7 @@ module.exports = {
           filepathPrefix: 'packages/apollo-client/src/',
         },
         sidebarCategories: {
-          null: [
-            'index',
-            'why-apollo',
-            'integrations',
-            'react-apollo-migration',
-          ],
+          null: ['index', 'why-apollo', 'integrations', 'hooks-migration'],
           Essentials: [
             'essentials/get-started',
             'essentials/queries',
@@ -58,9 +54,16 @@ module.exports = {
             'recipes/webpack',
             'recipes/meteor',
             'recipes/recompose',
-            'recipes/2.0-migration',
           ],
-          API: ['api/apollo-client', 'api/react-apollo'],
+          API: [
+            'api/apollo-client',
+            'api/react-hooks',
+            'api/react-ssr',
+            'api/react-testing',
+            'api/react-components',
+            'api/react-hoc',
+            'api/react-common',
+          ],
         },
       },
     },